+++ carpetmodSrc/carpet/logging/logHelpers/RNGMonitor.java
@@ -150,7 +150,11 @@ public class RNGMonitor {
     }
 
     public enum RNGAppType {
-        raw, fortune, mobSpawn, ironFarm, lightCheck, chunkTick, farmer
+        raw, fortune, mobSpawn, ironFarm,
+        /**
+         * Light Update Suppression
+         */
+        lightCheck, chunkTick, farmer
     }
 
     private static class RNGSeedBundle {